+++ spec/integration/resource/api300/c7000/fc_network/delete_spec.rb
@@ -2,13 +2,7 @@ require 'spec_helper'
 
 klass = OneviewSDK::API300::C7000::FCNetwork
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  include_context 'integration api300 context'
-
-  describe '#delete' do
-    it 'deletes the resource' do
-      item = OneviewSDK::API300::C7000::FCNetwork.new($client_300, name: FC_NET_NAME)
-      item.retrieve!
-      expect { item.delete }.not_to raise_error
-    end
+  include_examples 'FCNetworkDeleteExample', 'integration api300 context' do
+    let(:current_client) { $client_300 }
   end
 end