+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Features supported
 Added planned support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Ethernet network
+   - FC network
 
 # v4.2.0