+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## Suggested release: v4.3.0
 
 #### Features supported
-Added planned support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
+This release adds planned support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Ethernet network
    - FC network